Pyro AIR: Enable enumeration
Here I removed the to_event() calls so that Pyro won't turn Bernoulli into an Independent distribution (which doesn't support enumeration).
Independent distribution is the most useful when we need to denote event shape -- though in our case, z_pres only has shape of (1,), so we can just stick with Bernoulli (which assumes event shape is () ).
During sequential enumeration, z_pres will have value of (1,) (regardless of batch size), so we need to manually broadcast it back to the correct shape.
We might be able to enable parallel enumeration as well, though we have to figure out how to concatenate and pass the multiple values of z_pres into the neural modules.