iSplit sugar callee methods into `__call__` and `apply`.
One usability issue that remains in GenJAX is that `@dispatch` is used
for two implementations of `__call__` in SugaredGenerativeFunctionCall.
IDEs will type-check against one or the other, and typically pick the
`(key, *args)` form, since it's defined later, and appears to override
the `(*args, **kwargs)` form.

Having a `@dispatch` on __call__ thus "redlines" all the users's
function invoications.

I therefore propose that we take the less-common use, where you
manually supply the key and pass the args as a tuple, and rename
it `apply`, by analogy with LISP, where apply is used when you
already have your arguments in a list and don't want to splat
them out into a function call.