Update MaskingTrace.get_choices() to just return the inner choices.
I don't think it makes sense to Mask the choices -- the choices for
the inner generative function still need to be generated or specified,
whether the mask is True or False.

And, in the previous code, MaskingCombinator's assess, importance, and
update methods don't look at `choice.flag` for their `choice` argument.
(Because the mask must be specified by the input to the generative function,
not by its choice map.)

This change also makes it easier for users to construct/update choice maps
for masked generative functions (because they don't have to specify the mask
in the choice map).