use `not` instead of `jnp.logical_not` for CMS.has_addr
The use of logical_not produces a JNP object which results in a tracer object of type `bool[]`, and these cannot be used in `if has_addr` statements.

Effectively, I believe the use of logical_not changes the return type between original and complimented HierarchicalSelection objects, and so I don't think this is a case where we have introduced illegal control flow into something that is jitted.