Remove Sample and Constraint inheritance for ChoiceMap / Selection. Add `edit` & `generate` interfaces. (1/N) (#1322)
Then, fix the holes, and define overloads so users don't notice the
difference.



TODO:
- [ ] Can we change the signature of mask combinator in a follow up PR?
- [ ] Instead of a type assertion for checking equality between a
trace's generative function and self, could we check for tr.get_gen_fn()
is self ?

Here are the large scale changes which have occurred:
* `ChoiceMap` no longer inherits from both `Sample` and `Constraint` --
only from `Sample`
* `Selection` inherits from a new type called `Projection`, which is a
new concept -- denoting a way to break down a `Sample` into smaller
pieces (if allowed for that `Sample`).
* Introduced `ChoiceMapConstraint`, a type to cast a `ChoiceMap` into
`Constraint` (in general, old GenJAX users won't really need to care
about this type -- as we've added derived interfaces which retain the
interfaces where behaviors can be specified just via `ChoiceMap`,
without requiring wrapping in `ChoiceMapConstraint`)
* Added 3 new methods to generative functions: `generate`, `project`,
and `edit`. `generate` and `project` are actually specializations of
`edit`, but we keep them separate to make the implementation more
concise and readable -- to avoid introducing extra concepts like
`EmptyTrace`. These new methods are used to implement `importance` and
`update` -- whose signatures are specialized for people working with
generative functions whose sample space is `ChoiceMap`. The `generate`,
`project`, and `edit` methods are _more general_ than the derived
interfaces -- they are generic over the types `Constraint`,
`Projection`, and `EditRequest`. In other words, the old behavior people
are used to is retained -- but now as library authors, we can expand the
capabilities of the system by adding to the vocabulary of `edit` without
disrupting the derived interfaces.
* The interface `update` used to accept an `UpdateProblem` (this was
like v0 of `edit`) -- I've restored it to accept as signature `(key,
trace, choice_map, argdiffs)`.
* Given the above changes, I've propagated them throughout the code base
-- I've focused just on getting things with `ChoiceMap` working with the
above changes.
* I've made changes to the tests where signature changes caused errors
-- mostly with `update`.

I've added small comments in several of the files to point out specific
points of interest. There's a lot of spurious "I had to fix this,
because I changed this signature or this type" noise in this PR -- but
hopefully my summary above and the comments below allow a concise
summarized picture.